Remove intptr_t and use GTK's int/pointer casting

+++ src/panels.h
@@ -9,9 +9,9 @@
 
 void entry_edited(GtkWidget *entry, gpointer option);
 
-GtkWidget *create_toggle(const char *name, intptr_t id);
-GtkWidget *create_toggled_entry(const char *name, intptr_t id, gint width);
-GtkWidget *create_toggled_file(const char *name, intptr_t id, gint width,
+GtkWidget *create_toggle(const char *name, int id);
+GtkWidget *create_toggled_entry(const char *name, int id, gint width);
+GtkWidget *create_toggled_file(const char *name, int id, gint width,
 				const char *dialog_title);
-GtkWidget *create_toggled_combo(const char *name, intptr_t id, gint width,
+GtkWidget *create_toggled_combo(const char *name, int id, gint width,
 				GList *glist);
